update stack examples and deps
