--- example/custom/cache-control/.local-web-server.json
@@ -1,3 +1,0 @@
-{
-  "maxage": 2000
-}
