--- example/custom/live-reload/server.js
@@ -1,9 +1,0 @@
-'use strict'
-const Cli = require('../../')
-const liveReload = require('koa-livereload')
-
-const ws = new Cli()
-ws.addLogging('dev')
-  .add({ middleware: liveReload })
-  .addStatic()
-  .listen()
