+++ example/stack/cache-control.js
@@ -5,18 +5,22 @@ const DefaultStack = require('local-web-server-default-stack')
 
 class CacheControl extends DefaultStack {
   addAll () {
-    this.addLogging('dev')
-      .add({
-        optionDefinitions: {
-          name: 'maxage', type: Number,
-          description: 'The maxage to set on each response.'
-        },
-        middleware: function (options) {
-          return cacheControl({ maxAge: options.maxage })
-        }
-      })
+    return this.addLogging('dev')
+      .addCacheControl()
       .addStatic()
       .addIndex()
+  }
+  addCacheControl () {
+    this.add({
+      optionDefinitions: {
+        name: 'maxage', type: Number,
+        description: 'The maxage to set on each response.'
+      },
+      middleware: function (options) {
+        return cacheControl({ maxAge: options.maxage })
+      }
+    })
+    return this
   }
 }
