+++ example/stack/live-reload-optional/index.html
@@ -5,6 +5,6 @@
     <title>live-reload demo</title>
   </head>
   <body>
-    <h1>Live reloaded attached</h1>
+    <h1>Live reloaded potentially attached</h1>
   </body>
 </html>
