+++ lib/local-web-server.js
@@ -5,7 +5,6 @@ const path = require('path')
 const arrayify = require('array-back')
 const t = require('typical')
 const CommandLineTool = require('command-line-tool')
-const DefaultStack = require('local-web-server-default-stack')
 
 /**
  * @module local-web-server
@@ -30,7 +29,7 @@ class LocalWebServer {
       if (/^-/.test(stackPath)) stackPath = null
     }
 
-    const stackModule = loadStack(stackPath) || DefaultStack
+    const stackModule = loadStack(stackPath) || require('local-web-server-default-stack')
     this.stack = new stackModule()
     this.stack.addAll()
     const middlewareOptionDefinitions = this.stack.getOptionDefinitions()
