+++ package.json
@@ -34,38 +34,19 @@
   "dependencies": {
     "ansi-escape-sequences": "^2.2.2",
     "array-back": "^1.0.3",
-    "command-line-tool": "~0.3.0",
-    "config-master": "^2.0.2",
-    "http-proxy": "^1.13.3",
-    "kcors": "^1.2.1",
+    "command-line-tool": "~0.3.1",
+    "config-master": "^2.0.3",
     "koa": "^2.0.0",
-    "koa-bodyparser": "^3.0.0",
-    "koa-compose": "^3.1.0",
-    "koa-compress": "^1.0.9",
-    "koa-conditional-get": "^1.0.3",
-    "koa-convert": "^1.2.0",
-    "koa-etag": "^2.1.1",
-    "koa-json": "^1.1.3",
-    "koa-mock-response": "0.0.2",
-    "koa-morgan": "^1.0.1",
-    "koa-rewrite": "^2.1.0",
-    "koa-route": "^3.0.0",
-    "koa-send": "^3.2.0",
-    "koa-serve-index": "^1.1.1",
-    "koa-static": "^2.0.0",
-    "local-web-server-stack": "github:75lb/local-web-server-stack",
-    "path-to-regexp": "^1.5.0",
-    "reduce-flatten": "^1.0.0",
-    "stream-log-stats": "^1.1.3",
-    "test-value": "^2.0.0",
+    "local-web-server-default-stack": "github:local-web-server/default-stack",
+    "reduce-flatten": "^1.0.1",
     "typical": "^2.4.2",
     "walk-back": "^2.0.1"
   },
   "devDependencies": {
     "jsdoc-to-markdown": "^1.3.6",
     "koa-cache-control": "^1.0.0",
-    "koa-livereload": "^0.1.23",
+    "koa-livereload": "~0.2.0",
     "req-then": "~0.2.4",
-    "tape": "^4.5.1"
+    "tape": "^4.6.0"
   }
 }
